Finished main acquisition logic

+++ src/utilities/ca_code.rs
@@ -1,7 +1,7 @@
 use crate::constants::gps_ca_constants::GPS_CA_CODE_32_PRN;
 
 /// Generate CA code samples for 32 PRN code based on sampling frequency which might not be multiples of CA code rate
-pub fn generate_ca_code_samples(prn: usize, f_sampling: f32) -> Vec<i8> {
+pub fn generate_ca_code_samples(prn: u8, f_sampling: f32) -> Vec<i8> {
     let num_samples = (f_sampling
         / (gps_constants::GPS_L1_CA_CODE_RATE_CHIPS_PER_S
             / gps_constants::GPS_L1_CA_CODE_LENGTH_CHIPS))
@@ -13,7 +13,7 @@ pub fn generate_ca_code_samples(prn: usize, f_sampling: f32) -> Vec<i8> {
         })
         .collect();
 
-    let ca_code = GPS_CA_CODE_32_PRN[prn];
+    let ca_code = GPS_CA_CODE_32_PRN[prn as usize];
     
     samples_ind.iter().map(|&ind| ca_code[ind]).collect()
 }
